Set up some standard Ruby namespacing and file organization.
By simply dropping our procedure within the TicTacToe::Game module, we can
now interact with it from irb, or from any controller code that we'd like.  This
of course will be more useful when we add other functions to work with, but at
least makes it possible to load the code without running it immediately, and
moves the logic out of the global namespace.