Decompose board logic into a board object, clean up Game logic further.
Compared to our previous version, this code is much better structured.  If you
look at Board's implementation, you'll notice that all of its code works to
serve a single purpose: to implement a simple tic-tac-toe grid structure.  While
more verbose than what we had before, each individual feature is well isolated,
making it easier to test both manually and via unit tests.

If we revisit the Game class, we see that the flow is now very easy to follow,
and the main purpose of this code is now presentation and controller logic.  But
unlike Board which has a nice homogeneous feel to it, this code has a bit of a
leaky abstraction in that a few of its methods still look like they're
implementing some business logic.  While that's not something you'd need to
obsess over in ordinary day to day coding, for the purposes of this exercise,
let's try one more commit to address that point.