Break out rule logic into TicTacToe::Rules mixin.
This change causes TicTacToe::Game to become purely a controller class, with no
game specific logic within it.  Through the use of callbacks, TicTacToe::Rules
manages to be completely free of control code.

At this point, it is worth asking whether we've gone too far.  This design is
certainly more than what is required for such a trivial game, but the
implementation techniques we've used can certainly come in handy in more complex
scenarios.

One tricky thing about this level of abstraction is coming up with good names.
I'm not happy with 'check_move' as a name for a method that actually executes a
move, but it's a bit weird for it to be asymmetric to the check_win and
check_draw methods.  Perhaps the more generic move_action, win_action,
draw_action would have been better?  At this point, we need to say enough is
enough and call it a night.

Fun exercise though, right?